move unique for owner into validator folder
