+++ app/models/project.rb
@@ -55,7 +55,7 @@ class Project < ActiveRecord::Base
   validates_inclusion_of :private, :live, in: [true, false], message: "must be true or false"
 
   ## TODO: This potential has locking issues
-  validates_with UniqueForOwnerValidator
+  validates_with Validators::UniqueForOwnerValidator
 
   after_save :save_version
   after_update :send_notifications
