+++ spec/controllers/api/v1/collections_controller_spec.rb
@@ -214,7 +214,7 @@ describe Api::V1::CollectionsController, type: :controller do
       context 'when favorite collection already exists for given user/project' do
         before { create(:collection, favorite: true, owner: owner) }
 
-        it 'raises a 400 error on trying to create a favorite collection' do
+        it 'returns a bad request error trying to create a favorite collection' do
           post :create, favorite_params
           expect(response).to have_http_status(:bad_request)
         end
